Rename modules to include underbar.
